reshuffle files
